Putting pvt_logger into field_fluid_description. Updating the LICENSE file tol GPL v3
